Added setup file and broke apart functions into separate files
